End to End Data Science Project/3. Complete DS Project/requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
python-dotenv
mysql-connector-python
mysql
pymysql